pyproject.toml
[project]
name = "tariff_fetch"
version = "0.0.1"
description = "This is a repository for modern data science projects."
authors = [{ name = "Switchbox", email = "hello@switch.box" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastexcel>=0.16.0",
    "fuzzywuzzy>=0.18.0",
    "pathvalidate>=3.3.1",
    "polars>=1.34.0",
    "python-dotenv>=1.1.1",
    "python-levenshtein>=0.27.1",
    "questionary>=2.1.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "selenium>=4.36.0",
    "tenacity>=9.1.2",
    "typedload>=2.38",
    "typer>=0.19.2",
    "typing-extensions>=4.15.0",
]
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.urls]
Repository = "https://github.com/switchbox-data/tariff_fetch"
Documentation = "https://switchbox-data.github.io/tariff_fetch/"

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "ruff>=0.11.5",
    "ty>=0.0.1a21",
    "deptry>=0.20.0",
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.26.1",
    "build>=1.0.0",
    "twine>=4.0.0",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tariff_fetch"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
target-version = "py310"
line-length = 120
fix = true

[tool.ruff.lint]
select = ["E","W","F","B","I","UP","SIM","C4","RUF"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ty.environment]
python = "./.venv"
python-version = "3.13"
